BLFS script generation: use a boolean for stats, instead of a full test
Each time we have to decide whether some instructions should be added for
stats, we have to test whether the current sect1 id attribute
is contained in the list of packages for which stats are requested. This
used to be done with a full test each time. We now have a Boolean that we
pass to all the templates that need to use it.